requirements.txt
streamlit
numpy
scipy
matplotlib
pywebview